requirements.txt
elasticsearch==9.3.0
fastapi==0.129.2
langchain_core==1.2.14
langchain_milvus==0.3.3
langchain_openai==1.1.10
langgraph==1.0.9
pydantic==2.12.5
pymilvus==2.6.9
python-dotenv==1.2.1
Requests==2.32.5
rich==14.3.3
